pyproject.toml
[project]
name = "math-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=1.0.1",
    "langchain-core>=0.3.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.2.1",
]
